requirements.txt
# ──────────────────────────────────────────────
# Core Framework
# ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.7.4

# ──────────────────────────────────────────────
# HTTP Client (async, used for DuckDuckGo & Gemini)
# ──────────────────────────────────────────────
httpx==0.27.0

# ──────────────────────────────────────────────
# Environment variable loading
# ──────────────────────────────────────────────
python-dotenv==1.0.1
